Fix and organize a bunch of stuff
